fix: customer_idの型

+++ backend/src/main/java/server/resources/Customer.java
@@ -2,19 +2,19 @@ package server;
 
 public class Customer {
 
-	private final long id;
+	private final int id;
 	private final String name;
 	private final int creditLimit;
 	private final int creditTotal;
 
-	public Customer(long id, String name, int creditLimit, int creditTotal) {
+	public Customer(int id, String name, int creditLimit, int creditTotal) {
 		this.id = id;
 		this.name = name;
 		this.creditLimit = creditLimit;
 		this.creditTotal = creditTotal;
 	}
 
-	public long getId() {
+	public int getId() {
 		return id;
 	}
